+++ geostore/tests/test_views/test_schemas.py
@@ -9,7 +9,8 @@ from geostore.tests.factories import LayerFactory, UserFactory
 class SchemaValidationTest(TestCase):
     def setUp(self):
         self.client = APIClient()
-        self.user = UserFactory()
+        self.user = UserFactory(permissions=['geostore.can_manage_layers', ])
+
         self.client.force_authenticate(user=self.user)
 
         self.no_schema_layer = LayerFactory(name="no schema", geom_type=None)